MuonMDT_Cabling: Fix logging.
Need to cast int8_t types to int when writing to MsgStream; otherwise,
they're output as characters.

+++ MuonSpectrometer/MuonCablings/MuonMDT_Cabling/src/MuonMDT_CablingAlg.cxx
@@ -210,7 +210,7 @@ bool MuonMDT_CablingAlg::extractStationInfo(const coral::AttributeList& atr, Cab
         }
     }
     map_data.stationIndex = m_idHelperSvc->mdtIdHelper().stationNameIndex(stationNameString);
-    ATH_MSG_VERBOSE("station name: " << stationNameString << " index: " << map_data.stationIndex);
+    ATH_MSG_VERBOSE("station name: " << stationNameString << " index: " << static_cast<int>(map_data.stationIndex));
     // convert the subdetector id to integer
     map_data.subdetectorId = atoi(subdetector_id.c_str());
 
@@ -222,7 +222,7 @@ bool MuonMDT_CablingAlg::extractStationInfo(const coral::AttributeList& atr, Cab
     if (map_data.stationIndex < 0) {
         static std::atomic<bool> stWarningPrinted{false};
         if (!stWarningPrinted) {
-            ATH_MSG_WARNING("Found stationIndex=" << map_data.stationIndex
+            ATH_MSG_WARNING("Found stationIndex=" << static_cast<int>(map_data.stationIndex)
                                                   << " which is not reasonable, maybe related to ATLASRECTS-5961, continuing...");
             stWarningPrinted = true;
         }
